Modified the way to listen the notes
Removed the ability to hear notes on mouseover. It made everything a bit
too chaotic. Notes will be played now on mouse down. Moved the box
playbutton away from the middle so player can only listen the note of
the box if they want. This system also fits for mobile if it will ever
be done.